app/users/requirements.txt
asgiref==3.5.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
demoji==0.4.0
dj-database-url==0.5.0
Django==3.1
django-cleanup==5.1.0
django-crontab==0.7.1
django-filter==21.1
django-hosts==4.0
djangorestframework==3.13.1
dnspython==2.0.0
docopt==0.6.2
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
mailjet-rest==1.3.3
Paperboy==1.0.1
Pillow==9.0.1
pipreqs==0.4.10
psycopg2==2.9.3
py3-validate-email==0.2.10
-e git+https://github.com/gbozee/django-paystack.git@24df1c9ca209031e0ec940ff8552d45ca6b55421#egg=pypaystack
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tzdata==2021.5
tzlocal==2.1
urllib3==1.25.10
whitenoise==5.2.0
yarg==0.1.9
